[3.11] gh-78453: Move Unicode C API tests from test_unicode to test_capi.test_unicode (GH-99431). (cherry picked from commit 06d4e02c3b3526b5d90e41a0a0befa8663e08f27)
